Remove redundant casts

+++ src/main/java/me/pugabyte/nexus/features/particles/VectorUtils.java
@@ -105,8 +105,8 @@ public class VectorUtils {
 	}
 
 	public static Vector getBackVector(Location location) {
-		float newZ = (float) (location.getZ() + 1.0D * Math.sin(Math.toRadians((double) (location.getYaw() + 90.0F))));
-		float newX = (float) (location.getX() + 1.0D * Math.cos(Math.toRadians((double) (location.getYaw() + 90.0F))));
+		float newZ = (float) (location.getZ() + 1.0D * Math.sin(Math.toRadians(location.getYaw() + 90.0F)));
+		float newX = (float) (location.getX() + 1.0D * Math.cos(Math.toRadians(location.getYaw() + 90.0F)));
 		return new Vector((double) newX - location.getX(), 0.0D, (double) newZ - location.getZ());
 	}
 }
